added the ability to cud

+++ apps/onepiece/src/app/projects/projects.component.ts
@@ -44,9 +44,14 @@ export class ProjectsComponent implements OnInit {
     this.onepieceFacade.createOnepiece(onepiece)
   }
 
-  deleteOnepiece(onepiece) {
-    this.onepieceFacade.deleteOnepiece(onepiece)
-  }
+  // deleteOnepiece(onepiece) {
+    alert(onepiece) {
+      const confirmation = confirm(`Are you sure you want to delete ${onepiece.chapter}`)
+      if (confirmation) {
+        this.onepieceFacade.deleteOnepiece(onepiece);
+      }
+    }
+  // }
 
   cancel() {
     this.resetOnepiece();
